requirements.txt
Django>=3.1.1
djangorestframework>=3.12.1
numpy>=1.19.2
implicit>=0.4.4
django-cors-headers>=3.5.0
scipy>=1.5.2